# backend/scripts/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "BatchPredictionRequest": {
        "properties": {
          "events": {
            "description": "List of events to classify",
            "items": {
              "$ref": "#/components/schemas/EventData"
            },
            "title": "Events",
            "type": "array"
          },
          "user_context": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "User behavioral context",
            "title": "User Context"
          }
        },
        "required": [
          "events"
        ],
        "title": "BatchPredictionRequest",
        "type": "object"
      },
      "BatchPredictionResponse": {
        "properties": {
          "average_confidence": {
            "description": "Average confidence across all predictions",
            "title": "Average Confidence",
            "type": "number"
          },
          "predictions": {
            "description": "List of predictions",
            "items": {
              "$ref": "#/components/schemas/PriorityPredictionResponse"
            },
            "title": "Predictions",
            "type": "array"
          },
          "total_events": {
            "description": "Total number of events processed",
            "title": "Total Events",
            "type": "integer"
          }
        },
        "required": [
          "predictions",
          "total_events",
          "average_confidence"
        ],
        "title": "BatchPredictionResponse",
        "type": "object"
      },
      "Body_detect_scheduling_conflicts_api_v1_nlp_detect_conflicts_post": {
        "properties": {
          "new_event": {
            "$ref": "#/components/schemas/EventData"
          },
          "user_context": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Context"
          }
        },
        "required": [
          "new_event"
        ],
        "title": "Body_detect_scheduling_conflicts_api_v1_nlp_detect_conflicts_post",
        "type": "object"
      },
...
      "EventData": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Event description",
            "title": "Description"
          },
          "end_time": {
            "description": "Event end time (ISO format)",
            "title": "End Time",
            "type": "string"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Event location",
            "title": "Location"
          },
          "start_time": {
            "description": "Event start time (ISO format)",
            "title": "Start Time",
            "type": "string"
          },
          "title": {
            "description": "Event title",
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "start_time",
          "end_time"
        ],
        "title": "EventData",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PriorityExplanationResponse": {
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "feature_analysis": {
            "title": "Feature Analysis",
            "type": "object"
          },
          "priority": {
            "title": "Priority",
            "type": "integer"
          },
          "priority_label": {
            "title": "Priority Label",
            "type": "string"
          },
          "reasoning": {
            "title": "Reasoning",
            "type": "object"
          }
        },
        "required": [
          "priority",
          "priority_label",
          "confidence",
          "reasoning",
          "feature_analysis"
        ],
        "title": "PriorityExplanationResponse",
        "type": "object"
      },
      "PriorityPredictionRequest": {
        "properties": {
          "event": {
            "$ref": "#/components/schemas/EventData"
          },
          "user_context": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "User behavioral context",
            "title": "User Context"
          }
        },
        "required": [
          "event"
        ],
        "title": "PriorityPredictionRequest",
        "type": "object"
      },
      "PriorityPredictionResponse": {
        "properties": {
          "confidence": {
            "description": "Prediction confidence 0-1",
            "title": "Confidence",
            "type": "number"
          },
          "model_used": {
            "description": "Model type used for prediction",
            "title": "Model Used",
            "type": "string"
          },
          "priority": {
            "description": "Priority level 1-5",
            "title": "Priority",
            "type": "integer"
          },
          "priority_label": {
            "description": "Human-readable priority label",
            "title": "Priority Label",
            "type": "string"
          },
          "processing_time_ms": {
            "description": "Processing time in milliseconds",
            "title": "Processing Time Ms",
            "type": "number"
          }
        },
        "required": [
          "priority",
          "priority_label",
          "confidence",
          "model_used",
          "processing_time_ms"
        ],
        "title": "PriorityPredictionResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "AI-Powered Smart Calendar System with BERT Priority Classification",
    "title": "KairoCal API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/nlp/batch-predict": {
      "post": {
        "description": "Predict priorities for multiple events",
        "operationId": "batch_predict_priorities_api_v1_nlp_batch_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchPredictionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchPredictionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Batch Predict Priorities",
        "tags": [
          "NLP & BERT"
        ]
      }
    },
    "/api/v1/nlp/classify-priority": {
      "post": {
        "description": "Simple priority classification endpoint for direct event data\n\nArgs:\n    event_data: Dictionary containing event information\n    \nReturns:\n    Priority classification result with explanation",
        "operationId": "classify_event_priority_api_v1_nlp_classify_priority_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "Event Data",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Classify Event Priority",
        "tags": [
          "NLP & BERT"
        ]
      }
    },
...
    "/api/v1/nlp/detect-conflicts": {
      "post": {
        "description": "Detect conflicts for a new event using BERT-enhanced conflict detection",
        "operationId": "detect_scheduling_conflicts_api_v1_nlp_detect_conflicts_post",
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_detect_scheduling_conflicts_api_v1_nlp_detect_conflicts_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detect Scheduling Conflicts",
        "tags": [
          "NLP & BERT"
        ]
      }
    },
    "/api/v1/nlp/explain-priority": {
      "post": {
        "description": "Get detailed explanation for priority classification decision",
        "operationId": "explain_priority_decision_api_v1_nlp_explain_priority_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PriorityPredictionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriorityExplanationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Explain Priority Decision",
        "tags": [
          "NLP & BERT"
        ]
      }
    },
...
    "/api/v1/nlp/predict-priority": {
      "post": {
        "description": "Predict priority for a single event using BERT classifier",
        "operationId": "predict_event_priority_api_v1_nlp_predict_priority_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PriorityPredictionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriorityPredictionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Predict Event Priority",
        "tags": [
          "NLP & BERT"
        ]
      }
    },
...
  }
}